nk_backened/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pandas==2.2.2
openpyxl==3.1.2
python-multipart==0.0.9
pydantic==2.7.1
numpy==1.26.4
langchain
langchain-community
langchain-openai
faiss-cpu
pymupdf
streamlit
requests
python-dotenv
sentence-transformers
ollama
transformers 
accelerate 
sentence-transformers
finnhub-python